Using npm svgstore for more control on output
